+++ instrumentor/test/fixtures/is-pressable/output.js
@@ -1,0 +1,5 @@
+var MemoedPressable = (require('@heap/react-native-heap').default || {
+  withHeapPressableAutocapture: function withHeapPressableAutocapture(Component) {
+    return Component;
+  }
+}).withHeapPressableAutocapture(React.memo(React.forwardRef(Pressable)));